Add env variable to allow for local authentication
I just got this repo running locally and thought I'd make a few updates
in order to improve the dev experience working on the site locally and
for it to more closely match the production behavior.

This adds a new env variable `LOCAL_DEV_USER`
that when enabled allows the UI to behave like a logged in user. You can
use it like this:

```
LOCAL_RFD_REPO=~/code/rfd-dev LOCAL_DEV_USER=1 npm run dev
```

when starting up the dev server locally.

Now that we have logged-in and logged-out users locally this change
filters out RFDs based on their `visibility` flag instead of making them
ALL visible in your local RFD repo. This means if you are logged out
locally you won't see private RFDs. If you are logged in locally you
will be able to see private and public RFDs.

Also added some tests to document the behavior. I believe all of these
changes are safe to merge in (not affect prod) as they only modify the
local server file and for the changes in the `rfd.server.ts` file I only
modified code inside existing `if (isLocalMode()) {` blocks.