Add canonical URLs to fix duplicate page SEO issue
Went with `0053` as the canonical vs `53` since David said:

"We can decide whether the canonical URL is /rfd/0004 or /rfd/4. Not
sure there's any principled way to decide, but the former is what we use
on the site index currently."

I normally perfer super clean (and short) urls, but I think the
zero-padded version does preserve that "RFD 0053" aesthetic that's part
of the Oxide RFD culture so I went with that!

- Add <link rel="canonical"> tags to RFD pages, index, and login
- Redirect non-canonical RFD URLs (/rfd/53) to canonical format (/rfd/0053)
- Update internal links to use canonical format (formattedNumber)

Fixes #150